fix: Upgrade DoNotUseSizePropertyInAssert to avoid false positives (#87)
Currently, the "do not use .size on asserts" has several flaws and false
positives. A simple example is if `size` is called on something that is
not a map/collection, for which we have some suppressions on our main
codebase already.
This makes the rule much smarter, and also support new use cases, such
as within a `with` or other implicit `this` contexts.